Feature/#152: getDailyStreak()를 WidgetManager로 분리
