+++ HarubeeWidget/Sources/WidgetManager.swift
@@ -24,4 +24,48 @@ struct WidgetManager {
     let salaryBudget = try? salaryBudgetRepository.readByTargetDateContaining(.now)
     return salaryBudget
   }
+  
+  func getDailyStreak(_ salaryBudget: SalaryBudget?) -> [DailyStreak?] {
+    var dailyStreak: [DailyStreak?] = []
+    
+    guard let salaryBudget = salaryBudget,
+          let index = salaryBudget.dailyBudgets.firstIndex(where: {
+            $0.date == .now.formattedDate
+          }) else {
+      return Array(repeating: nil, count: 6)
+    }
+    
+    for i in 0..<6 {
+    
+      if index > salaryBudget.dailyBudgets.count - 1 {
+        dailyStreak.append(nil)
+        continue
+      }
+      
+      let dailyBudget = salaryBudget.dailyBudgets[index + i]
+      
+      let date = dailyBudget.date
+      let time: DailyStreak.Time = date == .now.formattedDate
+      ? .today : .future
+      let harubee = dailyBudget.harubee ?? Int(salaryBudget.defaultHarubee)
+      let isAdjustedHarubee = dailyBudget.harubee != nil ? true : false
+      let expenseType: DailyStreak.ExpenseType = {
+        guard let expense = dailyBudget.expense,
+              let income = dailyBudget.income else { return .empty }
+        
+        let result = harubee - expense + income
+        return result >= 0 ? .good : .bad
+      }()
+      
+      dailyStreak.append(DailyStreak(
+        date: date,
+        time: time,
+        harubee: harubee,
+        isAdjustedHarubee: isAdjustedHarubee,
+        expenseType: expenseType
+      ))
+    }
+    
+    return dailyStreak
+  }
 }
